refactor support routes
